Reorganiza pacotes dos serviços de yard e autenticação
