+++ backend/servico-autenticacao/src/main/java/br/com/cloudport/servicoautenticacao/config/SecurityFilter.java
@@ -1,6 +1,6 @@
 package br.com.cloudport.servicoautenticacao.config;
 
-import br.com.cloudport.servicoautenticacao.app.administracao.UserRepository;
+import br.com.cloudport.servicoautenticacao.repositories.UserRepository;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
