+++ backend/servico-autenticacao/src/main/java/br/com/cloudport/servicoautenticacao/config/UserInitializer.java
@@ -3,8 +3,8 @@ package br.com.cloudport.servicoautenticacao.config;
 import br.com.cloudport.servicoautenticacao.model.User;
 import br.com.cloudport.servicoautenticacao.model.UserRole;
 import br.com.cloudport.servicoautenticacao.model.Role;
-import br.com.cloudport.servicoautenticacao.app.administracao.UserRepository;
-import br.com.cloudport.servicoautenticacao.app.administracao.RoleRepository;
+import br.com.cloudport.servicoautenticacao.repositories.UserRepository;
+import br.com.cloudport.servicoautenticacao.repositories.RoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
