+++ backend/servico-autenticacao/src/main/java/br/com/cloudport/servicoautenticacao/controllers/AuthenticationController.java
@@ -1,4 +1,4 @@
-package br.com.cloudport.servicoautenticacao.app.configuracoes;
+package br.com.cloudport.servicoautenticacao.controllers;
 
 import br.com.cloudport.servicoautenticacao.app.configuracoes.dto.AuthenticationDTO;
 import br.com.cloudport.servicoautenticacao.app.configuracoes.dto.LoginResponseDTO;
@@ -7,10 +7,10 @@ import br.com.cloudport.servicoautenticacao.app.administracao.dto.UserInfoDTO;
 import br.com.cloudport.servicoautenticacao.model.Role;
 import br.com.cloudport.servicoautenticacao.model.User;
 import br.com.cloudport.servicoautenticacao.model.UserRole;
-import br.com.cloudport.servicoautenticacao.app.administracao.RoleRepository;
+import br.com.cloudport.servicoautenticacao.repositories.RoleRepository;
 import br.com.cloudport.servicoautenticacao.config.TokenService;
-import br.com.cloudport.servicoautenticacao.app.administracao.UserRepository;
-import br.com.cloudport.servicoautenticacao.app.administracao.UserRoleRepository;
+import br.com.cloudport.servicoautenticacao.repositories.UserRepository;
+import br.com.cloudport.servicoautenticacao.repositories.UserRoleRepository;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
