+++ backend/servico-autenticacao/src/main/java/br/com/cloudport/servicoautenticacao/controllers/RoleController.java
@@ -1,8 +1,7 @@
-package br.com.cloudport.servicoautenticacao.app.administracao;
+package br.com.cloudport.servicoautenticacao.controllers;
 
-import br.com.cloudport.servicoautenticacao.model.Role;
 import br.com.cloudport.servicoautenticacao.app.administracao.dto.RoleDTO;
-import br.com.cloudport.servicoautenticacao.app.administracao.RoleService;
+import br.com.cloudport.servicoautenticacao.services.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
