+++ backend/servico-autenticacao/src/main/java/br/com/cloudport/servicoautenticacao/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package br.com.cloudport.servicoautenticacao.app.administracao;
+package br.com.cloudport.servicoautenticacao.repositories;
 
 import br.com.cloudport.servicoautenticacao.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
